main_backend/requirements.txt
Flask==2.0.2
Flask-CORS==3.0.10
face_recognition==1.3.0
git+https://github.com/ageitgey/face_recognition_models
mysql-connector-python==8.0.26
numpy==1.21.2
opencv-python==4.5.3.56
